// packages/module-assignment/src/payload-types.ts
export interface User {
  id: number;
  firstName?: string | null;
  lastName?: string | null;
  role?: ('admin' | 'content-manager' | 'analytics-viewer' | 'instructor' | 'student') | null;
  bio?: string | null;
  theme: 'light' | 'dark';
  direction: 'ltr' | 'rtl';
  avatar?: (number | null) | Media;
  updatedAt: string;
  createdAt: string;
  enableAPIKey?: boolean | null;
  apiKey?: string | null;
  apiKeyIndex?: string | null;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  _verified?: boolean | null;
  _verificationToken?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
  collection: 'users';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  alt?: string | null;
  caption?: string | null;
  createdBy: number | User;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "courses".
 */
export interface Course {
  id: number;
  title: string;
  slug: string;
  description: string;
  descriptionMedia?: (number | Media)[] | null;
  status: 'draft' | 'published' | 'archived';
  recurringSchedules?:
    | {
        daysOfWeek: {
          day: number;
          id?: string | null;
        }[];
        startTime: string;
        endTime: string;
        startDate?: string | null;
        endDate?: string | null;
        id?: string | null;
      }[]
    | null;
  specificDates?:
    | {
        date: string;
        startTime: string;
        endTime: string;
        id?: string | null;
      }[]
    | null;
  thumbnail?: (number | null) | Media;
  tags?:
    | {
        tag?: string | null;
        id?: string | null;
      }[]
    | null;
  createdBy: number | User;
  sections?: {
    docs?: (number | CourseSection)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "course-sections".
 */
export interface CourseSection {
  id: number;
  course: number | Course;
  courseName?: string | null;
  courseSlug?: string | null;
  title: string;
  description?: string | null;
  parentSection?: (number | null) | CourseSection;
  parentSectionTitle?: string | null;
  contentOrder: number;
  childSections?: {
    docs?: (number | CourseSection)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
}